LinkedList done

+++ general/types.h
@@ -1,10 +1,12 @@
 #ifndef __TYPES_H__
 #define __TYPES_H__
 
+#include <cstddef>
+
 using T1 = int;
 using T  = unsigned int;
 using TT = unsigned long long;
-using Size = int;
+using Size = size_t;
 
 using ContainerElemType = int;
 using ContainerRange = int;
